Move converters outside integration
